+++ eis-jianming-master/src/main/java/com/prolog/eis/service/impl/unbound/calculation/CalculationTemplate.java
@@ -5,6 +5,7 @@ import com.prolog.eis.service.impl.unbound.entity.DetailDataBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Component
@@ -13,7 +14,7 @@ public class CalculationTemplate {
     @Autowired
     CalculationService calculationService;
 
-    public List<OutBoundSxStoreDto> calculationOutQty(DetailDataBean detailDataBeand, float miniPackage, boolean isPickStation){
+    public List<OutBoundSxStoreDto> calculationOutQty(DetailDataBean detailDataBeand, BigDecimal miniPackage, boolean isPickStation){
         return calculationService.buildOutBoundQty( detailDataBeand,  miniPackage,  isPickStation);
     }
